Ubah type data

+++ database/migrations/2025_07_21_002031_create_aspeks_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('aspeks', function (Blueprint $table) {
             $table->id();
             $table->string('nama');
-            $table->integer('weight');
+            $table->decimal('weight');
             $table->string('deskripsi');
             $table->timestamps();
         });
